chore(lib): avoid drawing markers with Plotly's `draw_paths` impl. (#258)
Previously, vertex markers were drawn with Plotly's implementation of `draw_paths` in the case were the number of vertices was small (< 20, according to their docs). This default behavior does not make much sense with respect to Matplotlib's and VisPy's implementation, and vertex markers will usually overlap with other markers such as the ones of TX/RX nodes. The default behavior is now to only draw lines, i.e., with `mode="lines"`.