feat(lib): deprecate `minimize` in favor of problem-specific solutions (#283)
Maintaining `differt.utils.minimize` is relatively complex, especially if we want to allow for arbitrary optimizers, objective functions, and batches.

As we are only using `minimize` once in the codebase (for Fermat path tracing), it makes more sense to deprecate the `minimize` function in favor of a specialized implementation. This reduces the complexity while hopefully allowing better performance.

If you were using `minimize` and don't know how to implement a `minimize` function yourself, feel free to ask for help in the discussions section.