feat(lib): add `TriangleMesh.mask` attribute
This should enable better execution time in code that requires generating meshes of different sizes, as using a mask allows to JIT-compile this function (when using a fill factor instead of a static size) and to stack multiple meshes into one. Masks should be transparent to users, and not impact the final output of Ray Tracing utilities.